:hammer: Restructure edits
