+++ src/Components/Cards/Image.tsx
@@ -3,12 +3,10 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { makeStyles, Theme } from '@material-ui/core/styles';
 import Dialog from '@material-ui/core/Dialog';
-import Grid from '@material-ui/core/Grid';
-import TextField from '@material-ui/core/TextField';
 
 import { CardBaseProps } from './CardBase';
 
-const useStyles = makeStyles((theme: Theme) => ({
+const useStyles = makeStyles((_theme: Theme) => ({
   root: {
     height: '100%'
   },
@@ -20,10 +18,6 @@ const useStyles = makeStyles((theme: Theme) => ({
     width: '100%',
     marginBottom: -6,
     cursor: 'pointer'
-  },
-  textField: {
-    width: 'calc(100% - 8px)',
-    margin: 4
   }
 }));
 
@@ -37,22 +31,6 @@ function Image(props: ImageProps) {
   }
 
   const classes = useStyles();
-
-  if (props.editing === 2)
-    return (
-      <Grid container direction="row" justify="center" alignItems="stretch">
-        <Grid item xs>
-        <TextField
-          className={classes.textField}
-          InputLabelProps={{ shrink: true }}
-          label="URL"
-          placeholder="http://timmo.dev/home-panel"
-          defaultValue={props.card.url}
-          onChange={props.handleChange!('url')}
-        />
-        </Grid>
-      </Grid>
-    );
 
   return (
     <div className={classes.root} onClick={handleToggleDialog}>
