+++ src/Components/Cards/Markdown.tsx
@@ -2,8 +2,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { makeStyles, Theme } from '@material-ui/core/styles';
-import TextField from '@material-ui/core/TextField';
-import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
 
 import { CardBaseProps } from './CardBase';
@@ -21,10 +19,6 @@ const useStyles = makeStyles((theme: Theme) => ({
         color: theme.palette.text.secondary
       }
     }
-  },
-  textField: {
-    width: 'calc(100% - 8px)',
-    margin: 4
   }
 }));
 
@@ -32,23 +26,6 @@ interface MarkdownProps extends CardBaseProps {}
 
 function Markdown(props: MarkdownProps) {
   const classes = useStyles();
-
-  if (props.editing === 2)
-    return (
-      <Grid container direction="row" justify="center" alignItems="stretch">
-        <Grid item xs>
-          <TextField
-            className={classes.textField}
-            InputLabelProps={{ shrink: true }}
-            multiline
-            label="Content"
-            placeholder="- Markdown"
-            defaultValue={props.card.content}
-            onChange={props.handleChange!('content')}
-          />
-        </Grid>
-      </Grid>
-    );
 
   return (
     <Typography
